Moved junit classes to the appropriate module
This way junit is not required in main apps.

resolves SCSH-28

Updated Deployer SPI
