+++ pom.xml
@@ -3,7 +3,7 @@
 	<modelVersion>4.0.0</modelVersion>
 
 	<artifactId>spring-cloud-scheduler-parent</artifactId>
-	<version>1.0.1.BUILD-SNAPSHOT</version>
+	<version>1.0.2.BUILD-SNAPSHOT</version>
 	<groupId>org.springframework.cloud</groupId>
 	<packaging>pom</packaging>
 
@@ -13,7 +13,7 @@
 	<parent>
 		<groupId>org.springframework.cloud</groupId>
 		<artifactId>spring-cloud-build</artifactId>
-		<version>1.3.10.RELEASE</version>
+		<version>1.3.11.RELEASE</version>
 		<relativePath />
 	</parent>
 
@@ -21,7 +21,7 @@
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<java.version>1.8</java.version>
 		<maven.version>3.5.0</maven.version>
-		<spring-cloud-deployer.version>1.3.3.RELEASE</spring-cloud-deployer.version>
+		<spring-cloud-deployer.version>1.3.4.RELEASE</spring-cloud-deployer.version>
 	</properties>
 
 	<modules>
